bot/picture_uploader/requirements.txt
boto3==1.21.5
python-dotenv==1.0.1
botocore==1.24.6
requests==2.26.0
aiohttp==3.9.1